Integrate fancy-rocc.py into ROCC-coffea.py
